+++ current/src/2f/Listeners/listener.cpp
@@ -36,7 +36,7 @@ void F2::Listener::onInit()
 	
 }
 
-void F2::Listener::onTicked(float const& tick)
+void F2::Listener::onTicked(int const& tick)
 {
 	
 }
